This checkin corrects the previous checkin. The error was the following:
We are given a collection [P1;P2;...;Pn] of collections of points (1x3 vectors)
stored in objectivefn_data (and the number of points in Pi is stored in
objectivefn_partition(1,i)). We want to compute each 3-tuple of points
in Pi, i=1,...,n; this is Pi^{3}, the set of 3-subsets of Pi. Then, we
want, for each pair of planes Qi, Qj in Pi^{3}, Pj^{3}, we compute a
line Lij = intersection of Qi with Qj (i!=j).

So, we want to iterate over all 2-subsets (i,j) of 1:n, then iterate over all
3-subsets of Pi^{3} and Pj^{3} to compute all possible lines Lij, which is
stored in objectivefn_lines.

Previously, we iterated over all 3-subsets, then all 2-subsets, which leads to
duplications.

As an example, take a partition
[4,4,4;3,3,3]

We have 4 x 4 x 4=4^3 3-subsets. From this partition, we get
4x4+4x4+4x4=48 lines.

On the other hand, if we enumerate all 3-subsets, then all 2-subsets, we get
4^3 choose 2 = 2016 lines.